requirements.txt
googlemaps
python-dotenv
pydantic-settings
pydantic
fastapi
uvicorn
mecab-python3
unidic-lite
geocoder